requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
jinja2>=3.1.3
python-multipart>=0.0.7
bcrypt>=4.1.2
openpyxl>=3.1.2
python-dotenv>=1.0.1
reportlab>=4.0.0
qrcode[pillow]>=7.4.2
docxtpl>=0.16.0
python-docx>=1.1.0
